优化简历编辑页接口

+++ src/api/resume/index.ts
@@ -1,18 +1,24 @@
 import type { ResumeResponse } from "@/@type";
 
 import http from "@/utils/http";
+import { getItem } from "@/utils/storage";
 import { nanoid } from "@reduxjs/toolkit";
 
 /**
  * 创建简历的接口
  */
-export const createResume = (userId: string, value: string, theme: string) => {
+export const createResume = (value: string, theme: string) => {
   const id = nanoid();
+  const username = getItem("username");
+  if (!username) {
+    return Promise.resolve(null);
+  }
+
   return http.request<ResumeResponse>({
-    url: `/home/addresume?userId=${userId}`,
+    url: `/home/addresume?userId=${username}`,
     method: "post",
     data: {
-      userId: userId,
+      userId: username,
       resumeText: value,
       resumeName: `template-${id}`,
       resumeTheme: theme,
@@ -23,10 +29,13 @@ export const createResume = (userId: string, value: string, theme: string) => {
 /**
  * 获取所有简历的接口
  */
-export const getResumeAll = (userId: string) => {
-  // TODO
+export const getResumeAll = () => {
+  const username = getItem("username");
+  if (!username) {
+    return Promise.resolve(null);
+  }
   return http.request<Array<ResumeResponse>>({
-    url: `/home/getresume?userId=${userId}`,
+    url: `/home/getresume?userId=${username}`,
     method: "get",
   });
 };
@@ -34,10 +43,13 @@ export const getResumeAll = (userId: string) => {
 /**
  * 获取简历的接口
  */
-export const getResume = (userId: string, resumeId: string) => {
-  // TODO
+export const getResume = (resumeId: string) => {
+  const username = getItem("username");
+  if (!username) {
+    return Promise.resolve(null);
+  }
   return http.request<ResumeResponse>({
-    url: `/home/getresumeone?userId=${userId}&id=${resumeId}`,
+    url: `/home/getresumeone?userId=${username}&id=${resumeId}`,
     method: "get",
   });
 };
@@ -45,23 +57,30 @@ export const getResume = (userId: string, resumeId: string) => {
 /**
  * 保存简历的接口
  */
-export const modifyResume = (userId: string, value: string, id: string) => {
-  // TODO
+export const modifyResume = (value: string, id: string) => {
+  const username = getItem("username");
+  if (!username) {
+    return Promise.resolve(null);
+  }
   return http.request<ResumeResponse>({
-    url: `/home/updateresume?userId=${userId}&id=${id}`,
+    url: `/home/updateresume?userId=${username}&id=${id}`,
     method: "put",
     data: {
       id: id,
-      userId: userId,
+      userId: username,
       resumeText: value,
       resumeName: `template-${id}`,
     },
   });
 };
 
-export const deleteResume = (userId: string, resumeId: string) => {
+export const deleteResume = (resumeId: string) => {
+  const username = getItem("username");
+  if (!username) {
+    return Promise.resolve(null);
+  }
   return http.request<null>({
-    url: `/home/deleteresume?userId=${userId}&id=${resumeId}`,
+    url: `/home/deleteresume?userId=${username}&id=${resumeId}`,
     method: "delete",
   });
 };
